KBX-412 Render the K-7 module inside the cabinet layout
The K-7 pages lived in their own (k-reports) route group with a standalone
header, so opening the module dropped the sidebar and looked like a separate
application. The pages now sit in the (private) group, wrapped in SubLayout
with breadcrumbs like every other cabinet module; the employee-only guard
moved to a nested layout.
