+++ src/app/[locale]/(private)/k-7/components/k7-report-error-dialog.tsx
