+++ src/app/[locale]/(private)/k-7/components/k7-reports-table.tsx
