+++ src/app/[locale]/(private)/k-7/preview/components/k7-preview-loading.tsx
