+++ src/app/[locale]/(private)/k-7/preview/components/methodical-work-table.tsx
