+++ src/app/[locale]/(private)/k-7/preview/components/other-duties-table.tsx
