+++ src/app/[locale]/(private)/k-7/preview/components/report-section.tsx
