+++ src/app/[locale]/(private)/k-7/preview/components/scientific-work-table.tsx
