+++ src/app/[locale]/(private)/k-7/preview/components/summary-table.tsx
