+++ src/app/[locale]/(private)/k-7/preview/components/table-styles.ts
