+++ src/app/[locale]/(private)/k-7/preview/components/teaching-disciplines-table.tsx
