+++ src/app/[locale]/(private)/k-7/preview/constants.ts
