+++ src/app/[locale]/(private)/k-7/preview/page.content.tsx
